u-boot: Add peach pit chromebook configs
This adds U-Boot configuration for the Samsung Chromebook 2 11", also
known as "peach-pit" in the U-Boot upstream defconfigs. Also adds a
shared "peach" board directory to share with others having the same
baseboard. The config is almost the same as upstream defconfig, but with
REMAKE_ELF and POSITION_INDEPENDENT enabled.

Untested since I don't have the peach pit chromebook.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>