u-boot: Add peach pi chromebook configs
This adds U-Boot configuration for the Samsung Chromebook 2 13", also
known as "peach-pi" in the U-Boot upstream defconfigs. It uses the
shared tree for the "peach" baseboard. The config is almost the same as
upstream defconfig, but with REMAKE_ELF and POSITION_INDEPENDENT
enabled.

Untested since I don't have the peach pi chromebook. Note the there
doesn't seem to be any coreboot support for this chromebook.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>