u-boot: Add snow chromebook configs
This adds U-Boot configuration for the Samsung Chromebook - XE303, also
known as "snow" in the U-Boot upstream defconfigs. Also adds a shared
"daisy" board directory to share with others having the same baseboard.
The config is almost the same as upstream defconfig, but with REMAKE_ELF
and POSITION_INDEPENDENT enabled.

Untested since I don't have the snow chromebook.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>