u-boot: Add spring chromebook configs
This adds U-Boot configuration for the HP Chromebook 11 G1, also known
as "spring" in the U-Boot upstream defconfigs. It uses the shared tree
for the "daisy" baseboard. The config is almost the same as upstream
defconfig, but with REMAKE_ELF and POSITION_INDEPENDENT enabled.

Untested since I don't have the spring chromebook.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>