delete the Makefile
we don't use it in lbmk. it's there mostly because
it was technically feasible, and it still is

however, i've been doing massive re-factoring of
lbmk and the makefile and i just don't feel like
constantly patching up the makefile

if someone wants to re-add it, that's fine. but i
don't see the point in maintaining something that
we don't need.

the makefile is not needed. all it did was call
lbmk directly. the makefile had no logic itself.

Signed-off-by: Leah Rowe <leah@libreboot.org>